# Refactor detect_clearsky