… benchmarks/benchmarks/detect_clearsky.py
… docs/sphinx/source/whatsnew/v0.8.1.rst
… pvlib/data/detect_clearsky_data.csv
… pvlib/tests/test_clearsky.py