- [x] I am familiar with the [contributing guidelines](https://pvlib-python.readthedocs.io/en/latest/contributing.html)
 - [x] Tests added
 - [x] Adds description and name entries in the appropriate "what's new" file in [`docs/sphinx/source/whatsnew`](https://github.com/pvlib/pvlib-python/tree/master/docs/sphinx/source/whatsnew) for all changes. Includes link to the GitHub Issue with `` :issue:`num` `` or this Pull Request with `` :pull:`num` ``. Includes contributor name and/or GitHub username (link with `` :ghuser:`user` ``).
 - [x] New code is fully documented. Includes [numpydoc](https://numpydoc.readthedocs.io/en/latest/format.html) compliant docstrings, examples, and comments where necessary.
 - [x] Pull request is nearly complete and ready for detailed review.
 - [x] Maintainer: Appropriate GitHub Labels and Milestone are assigned to the Pull Request and linked Issue.

Refactor of `clearsky.detect_clearsky` in preparation for adding BrightSun detect clearsky algorithm. Major changes:
- uses pd.Series.rolling instead of Hankel matrix to compute criteria over sliding windows. This change accomodates centered windows, i.e., aligning the criteria for a window to a midpoint rather than to the left endpoint (default behavior in this code).
- calculation of criteria delegated to helper functions.
- new kwarg `align` (default 'left') otherwise function signature is unchanged.