@cwhanse 2020-10-05
if `clearsky.detect_clearsky` is of interest please take a look, in particular, at `_calc_stats` and the other helper functions.

I need to add some tests for the non-trivial helper functions, and I'm sure there's room for performance improvement. The `align` feature is added anticipating the BrightSun method which uses centered windows. I think that the BrightSun method could be implemented using numpy arrays and the Hankel matrix, but using pandas.Series.rolling seems more transparent to me.

We could drop `align='right'` as an option (the current code's behavior is `align='left'` but `pandas.rolling` only offers `center=False (default, means right-aligned) and `center=True` (centered). Maintaining the current behavior with `align='left'` by default wasn't too hard. I'm not sure how much value the `align` options offer, but there they are. `align='center'` is more in line (sorry) with what Matt and I had in mind when we wrote the algorithm, but we didn't code it that way in Matlab.