@wholmgren 2020-10-06
Is there a compelling use case for `align='left'`? It sounds to me like we should consider it a bug in the initial implementation, and always use `center=True` (without exposing a kwarg). 

I will carefully review when I have the time. I want to understand why this implementation works but the previous attempts to use pandas did not work.