@cwhanse 2020-10-06
`align='left'` is here to preserve the old behavior, personally I agree with `align='center'` as the better choice for a default.

I also think we could drop support for `array` inputs and lose the `times` argument.