@wholmgren 2020-10-06
To put it differently: why should we preserve the old behavior? It sounds like a bug. People that want the old behavior can use an old, buggier version of pvlib.

> I also think we could drop support for array inputs and lose the times argument.

agree