@cwhanse 2020-10-08
> Other than consistency with previous versions, when might a user prefer 'left' or 'right'? 

Consistency is the only reason I came up with. The option to align 'left', 'center', 'right' falls out of the combination of numpy and pandas defaults.

> I'd hard-code center 

I'm OK with this.