@cwhanse 2020-12-15
> To put it differently: why should we preserve the old behavior? It sounds like a bug. People that want the old behavior can use an old, buggier version of pvlib.
> 
> > I also think we could drop support for array inputs and lose the times argument.
> 
> agree

@wholmgren prefer to deprecate removing the `times` argument and restricting input type to Series?