@wholmgren 2020-12-15
Ok with me but how do we gracefully deprecate a required argument from the middle of a signature?