@cwhanse 2020-12-15
> > @wholmgren prefer to deprecate removing the `times` argument and restricting input type to Series?
> 
> Ok with me but how do we gracefully deprecate a required argument from the middle of a signature? I think in the past we've only done this kind of change when moving a function to a new name or module.


"gracefully" idk, maybe accept array types and the `times` argument, and use the `times` argument but raise a warning if it is supplied.

Or release it as a breaking API change.

My guess is that the array-type users are in the minority and the Series-type users would appreciate losing the `times` parameter.