@wholmgren 2020-12-15
I'm not opposed to delaying this to 0.9 and making it a breaking change. We can do a non-breaking change if we change both `times` and `window_length` to kwargs. Maybe someone else has a better idea.