@cwhanse 2020-12-16
Let's leave the parameters unchanged. In v0.9 I'd want to rename this function to `detect_clearsky_reno` to make room for `detect_clearsky_brightsun`. We can adjust the parameters in the renamed function.