@cwhanse 2020-12-22
> * Consider dropping the newly unnecessary `times` argument on line 643 in `clearsky.rst`.

We decided not to drop the 'times' argument in this refactor, but rather to rename this function in v0.9 to `detect_clearsky_reno` and drop the 'times' argument then.

> * I notice that the reference performed the rescaling "using only clear periods on days when greater than 50% of the day was classified as clear".  Just pointing it out in case it's worth doing the same here or documenting the difference.

Adding a comment to this effect.