@kanderso-nrel 2020-12-22
> 
> 
> > * Consider dropping the newly unnecessary `times` argument on line 643 in `clearsky.rst`.
> 
> We decided not to drop the 'times' argument in this refactor, but rather to rename this function in v0.9 to `detect_clearsky_reno` and drop the 'times' argument then.

I'm on board with that plan for 0.9 -- to clarify, this suggestion was about this [documentation page](https://pvlib-python.readthedocs.io/en/stable/clearsky.html#detect-clearsky) where the inputs are Series and passing `cs.index` isn't needed anymore.  OK with me to just leave it until 0.9 when the parameter goes away for good.