@wholmgren 2020-12-23
https://github.com/cwhanse/pvlib-python/pull/6 adds an asv test for this function. Results on my machine for the latest commit 5927d73 using `asv run -b 'detect_clearsky.DetectClear.time_detect_clearsky' HEAD`

```
[  0.14%] · For pvlib-python commit 5927d738 <pr/cwhanse/1074>:
[  0.14%] ·· Building for conda-py3.6-ephem3.7.6.0-numba0.36.1-numpy1.12.0-pandas0.22.0-pytables3.6.1-scipy1.2.0....
[  0.14%] ·· Benchmarking conda-py3.6-ephem3.7.6.0-numba0.36.1-numpy1.12.0-pandas0.22.0-pytables3.6.1-scipy1.2.0
[  0.17%] ··· Running (detect_clearsky.DetectClear.time_detect_clearsky--).
[  0.21%] ··· detect_clearsky.DetectClear.time_detect_clearsky                                                                                                                                                                      1.27±0.02s
[  0.21%] ·· Building for conda-py3.8-ephem-numba-numpy-pandas-pytables-scipy....
[  0.21%] ·· Benchmarking conda-py3.8-ephem-numba-numpy-pandas-pytables-scipy
[  0.24%] ··· Running (detect_clearsky.DetectClear.time_detect_clearsky--).
[  0.28%] ··· detect_clearsky.DetectClear.time_detect_clearsky                                                                                                                                                                      13.5±0.03s
```

disconcerting result for the python 3.8 environment.