@cwhanse 2020-12-23
thanks @wholmgren  I profiled the code and nearly all the time is being spent in those two '.apply' methods. That's the good news, I know where to focus. I don't think this should be merged with such a drastic loss of performance.