@kanderso-nrel 2020-12-23
I would guess that the order of magnitude slowdown between the two asv environments is from pandas introducing the optional `raw` parameter in v0.23.0, making `Rolling.apply()` switch from passing arrays to passing Series by default.