@wholmgren 2020-12-23
> I would guess that the order of magnitude slowdown between the two asv environments is from pandas introducing the optional raw parameter in v0.23.0, making Rolling.apply() switch from passing arrays to passing Series by default.

That's a good guess. If true, we should bump the min pandas requirement so that we can pass `raw=True` without concern for the version. 

@cwhanse I'm fine with holding up 0.8.1 for a little bit while you experiment. I just want it released by the time we'd want to merge the Array PR.