@cwhanse 2020-12-23
I looked at passing 'raw' (for later pandas versions) and didn't see a speedup. I could have done it wrong. I think the issue is that if the function being applied isn't in the set of precompiled methods, that `.apply` loops through the windows.  This changes from using `pd.rolling.apply` to `np.apply_along_axis` which is ~x10 faster.