@wholmgren 2020-12-23
```
asv compare 6e936f5 HEAD                                                                                                                                    1 ↵

All benchmarks:

       before           after         ratio
     [6e936f57]       [c3c1f9e3]
     <pr/cwhanse/1074~5^2>       <detect_clearsky>
+      14.5±0.3ms       2.26±0.04s   155.72  detect_clearsky.DetectClear.time_detect_clearsky [holmgren-mbp.local/conda-py3.6-ephem3.7.6.0-numba0.36.1-numpy1.12.0-pandas0.22.0-pytables3.6.1-scipy1.2.0]
+      16.1±0.4ms       2.30±0.06s   142.84  detect_clearsky.DetectClear.time_detect_clearsky [holmgren-mbp.local/conda-py3.8-ephem-numba-numpy-pandas-pytables-scipy]
```