@wholmgren 2020-12-24
```
asv compare 6e936f5 HEAD

All benchmarks:

       before           after         ratio
     [6e936f57]       [983bd6f8]
     <pr/cwhanse/1074~5^2>       <detect_clearsky>
+      14.5±0.3ms       55.6±0.8ms     3.83  detect_clearsky.DetectClear.time_detect_clearsky [holmgren-mbp.local/conda-py3.6-ephem3.7.6.0-numba0.36.1-numpy1.12.0-pandas0.22.0-pytables3.6.1-scipy1.2.0]
+      16.1±0.4ms         54.2±2ms     3.36  detect_clearsky.DetectClear.time_detect_clearsky [holmgren-mbp.local/conda-py3.8-ephem-numba-numpy-pandas-pytables-scipy]
```

I'm willing to accept this performance decrease if we think it's a net win for readability and ease of reuse in future algorithms.